+++ CHANGELOG.md
@@ -48,6 +48,10 @@
 ## [Known Issues] - <em>Send bug reports in More > Send Feedback</em><br><em>Only issues in the stable version will be listed here</em>
 - None
 
+## [Stable 6.0.6] - 2023-05-23
+### Fixed
+- Fix locked PowerSchool sync
+
 ## [Announcement 6.0.5] - 2023-05-12 - <span class="changelog-inline-setoff" style="color:red">CLICK ON THIS. READ IT.</span>
 - <span class="changelog-inline-setoff">TODAY, MAY 12, 2023</span>
 - <span class="changelog-inline-setoff">The Developers are in front of Sobrato</span>